Refactor Comparator class
In the previous version of the Comparator class, the returned List included lines of data where both value1 from the first file and value2 from the second file were presented, even if there was no such value in the corresponding file (in that case, value == null).

Now the returned List only contains lines with usable data. If a file doesn't contain a specific value, it won't appear in the line of data.